refactor: move all evaluation code into evaluation folder
