+++ CMakeLists.txt
@@ -12,7 +12,6 @@ set(COMMON_SOURCES
     src/board/playmove.cpp
     src/board.cpp
     src/evaluation/evaluation.cpp
-    src/evaluation/psqt.cpp
     src/move_gen/king.cpp
     src/move_gen/pawn.cpp
     src/move_gen.cpp
