+++ src/board/playmove.cpp
@@ -2,6 +2,7 @@
 
 #include <stdexcept>
 
+#include "evaluation/evaluation.hpp"
 #include "utils.hpp"
