+++ src/piece.hpp
@@ -1,13 +1,6 @@
 #ifndef PIECE_HPP
 #define PIECE_HPP
 
-#define KING_VALUE 100000
-#define QUEEN_VALUE 900
-#define ROOK_VALUE 500
-#define BISHOP_VALUE 300
-#define KNIGHT_VALUE 300
-#define PAWN_VALUE 100
-
 #include "defs.hpp"
